+++ src/main/java/com/armaandhir/demo/service/UserAccountService.java
@@ -1,19 +1,40 @@
 package com.armaandhir.demo.service;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import com.armaandhir.demo.model.UserAccount;
 
+/**
+ * @author Armaan Dhir
+ *
+ */
 public interface UserAccountService {
 
+	/**
+	 * @return
+	 */
 	Collection<UserAccount> findAll();
 	
-	UserAccount findOne(BigInteger id);
+	/**
+	 * @param id
+	 * @return
+	 */
+	UserAccount findOne(Long id);
 	
+	/**
+	 * @param account
+	 * @return
+	 */
 	UserAccount create(UserAccount account);
 	
+	/**
+	 * @param account
+	 * @return
+	 */
 	UserAccount update (UserAccount account);
 	
-	void delete (BigInteger id);
+	/**
+	 * @param id
+	 */
+	void delete (Long id);
 }
